seperate cli.ts to multiple files for each command.
