+++ src/cli/error.ts
@@ -1,0 +1,4 @@
+export const error = (message: any) => {
+    console.error(message.toString())
+    process.exit(1)
+}
