+++ tsconfig.json
@@ -15,5 +15,5 @@
     "module": "commonjs",
     "outDir": "./dist"
   },
-  "include": ["src/cli.ts"]
+  "include": ["src/cli/**.ts"]
 }
